chore: move WrapAddress in utils
